+++ CHANGELOG.md
@@ -9,6 +9,18 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleased]
+
+### Added
+
+- Add `mysql.IsCompatibleWith()`
+- Add `postgres.IsCompatibleWith()`
+- Add `sqlite.IsCompatibleWith()`
+
+### Fixed
+
+- Fix `driver.New()` to build correctly when CGO is disabled
+
 ## [0.2.0] - 2019-10-19
 
 ### Added